+++ axum/src/response/mod.rs
@@ -15,6 +15,9 @@ pub use crate::Json;
 #[cfg(feature = "headers")]
 pub use crate::TypedHeader;
 
+#[doc(no_inline)]
+pub use crate::Extension;
+
 #[doc(inline)]
 pub use axum_core::response::{IntoResponse, IntoResponseParts, Response, ResponseParts};